Crate revm_state

Optimism-specific constants, types, and helpers.

Re-exports

pub use bytecode;

pub use primitives;

Structs

Account

Account type used inside Journal to track changed to state.

AccountInfo

Account information that contains balance, nonce, code hash and code

AccountStatus

Account status flags. Generated by bitflags crate.

EvmStorageSlot

This type keeps track of the current value of a storage slot.

Enums

Bytecode

Main bytecode structure with all variants.

Type Aliases

EvmState

EVM State is a mapping from addresses to accounts.

EvmStorage

An account’s Storage is a mapping from 256-bit integer keys to EvmStorageSlots.

TransientStorage

Structure used for EIP-1153 transient storage