revm_state

Struct AccountInfo

pub struct AccountInfo {
    pub balance: U256,
    pub nonce: u64,
    pub code_hash: B256,
    pub code: Option<Bytecode>,
}

Account information that contains balance, nonce, code hash and code

Code is set as optional.

Fields
balance: U256

Account balance.

nonce: u64

Account nonce.

code_hash: B256

Hash of the raw bytes in code, or KECCAK_EMPTY.

code: Option<Bytecode>

Bytecode data associated with this account.

If None, code_hash will be used to fetch it from the database, if code needs to be loaded from inside revm.

By default, this is Some(Bytecode::default()).

Implementations

impl AccountInfo

pub fn new(balance: U256, nonce: u64, code_hash: B256, code: Bytecode) -> Self

Creates a new AccountInfo with the given fields.

pub fn with_code(self, code: Bytecode) -> Self

Creates a new AccountInfo with the given code.

Note

As code hash is calculated with Bytecode::hash_slow there will be performance penalty if used frequently.

pub fn with_code_hash(self, code_hash: B256) -> Self

Creates a new AccountInfo with the given code hash.

Note

Resets code to None. Not guaranteed to maintain invariant code and code_hash. See also Self::with_code_and_hash.

pub fn with_code_and_hash(self, code: Bytecode, code_hash: B256) -> Self

Creates a new AccountInfo with the given code and code hash.

Note

In debug mode panics if Bytecode::hash_slow called on code is not equivalent to code_hash. See also Self::with_code.

pub fn with_balance(self, balance: U256) -> Self

Creates a new AccountInfo with the given balance.

pub fn with_nonce(self, nonce: u64) -> Self

Creates a new AccountInfo with the given nonce.

pub fn set_balance(&mut self, balance: U256) -> &mut Self

Sets the AccountInfo balance.

pub fn set_nonce(&mut self, nonce: u64) -> &mut Self

Sets the AccountInfo nonce.

pub fn set_code_hash(&mut self, code_hash: B256) -> &mut Self

Sets the AccountInfo code_hash and clears any cached bytecode.

Note

Calling this after set_code(...) will remove the bytecode you just set. If you intend to mutate the code, use only set_code.

pub fn set_code(&mut self, code: Bytecode) -> &mut Self

Replaces the AccountInfo bytecode and recalculates code_hash.

Note

As code hash is calculated with Bytecode::hash_slow there will be performance penalty if used frequently.

pub fn set_code_and_hash(&mut self, code: Bytecode, code_hash: B256)

Sets the bytecode and its hash.

Note

It is on the caller’s responsibility to ensure that the bytecode hash is correct.

pub fn copy_without_code(&self) -> Self

Returns a copy of this account with the Bytecode removed.

This is useful when creating journals or snapshots of the state, where it is desirable to store the code blobs elsewhere.

Note

This is distinct from without_code in that it returns a new AccountInfo instance with the code removed.

without_code will modify and return the same instance.

pub fn without_code(self) -> Self

Strips the Bytecode from this account and drop it.

This is useful when creating journals or snapshots of the state, where it is desirable to store the code blobs elsewhere.

Note

This is distinct from copy_without_code in that it modifies the account in place.

copy_without_code will copy the non-code fields and return a new AccountInfo instance.

pub fn is_empty(&self) -> bool

Returns if an account is empty.

An account is empty if the following conditions are met.

code hash is zero or set to the Keccak256 hash of the empty string ""

balance is zero

nonce is zero

pub fn exists(&self) -> bool

Returns true if the account is not empty.

pub fn has_no_code_and_nonce(&self) -> bool

Returns true if account has no nonce and code.

pub fn code_hash(&self) -> B256

Returns bytecode hash associated with this account.

If account does not have code, it returns KECCAK_EMPTY hash.

pub fn is_empty_code_hash(&self) -> bool

Returns true if the code hash is the Keccak256 hash of the empty string "".

pub fn take_bytecode(&mut self) -> Option<Bytecode>

Takes bytecode from account.

Code will be set to None.

pub fn from_balance(balance: U256) -> Self

Initializes an AccountInfo with the given balance, setting all other fields to their default values.

pub fn from_bytecode(bytecode: Bytecode) -> Self

Initializes an AccountInfo with the given bytecode, setting its balance to zero, its nonce to 1, and calculating the code hash from the given bytecode.